feat(core): move to separate agent directory
metriport/metriport-internal#1040
