+++ packages/core/src/external/bedrock/agent/types.ts
@@ -1,0 +1,8 @@
+import type { BedrockTool } from "./tool";
+
+export interface BedrockAgentConfig {
+  systemPrompt: string;
+  tools?: BedrockTool[];
+  maxTokens?: number;
+  temperature?: number;
+}
